old_project/requirements.txt
alembic==0.9.9
APScheduler==3.5.1
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-APScheduler==1.8.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
lxml==3.8.0
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
PyMySQL==0.8.0
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.4
tzlocal==1.5.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
